pyproject.toml
[project]
name = "randomfastaseqgen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.85",
    "pandas>=2.2.3",
    "pyqt6>=6.9.0",
    "seaborn>=0.13.2",
]
